Fix hdf5 formats
- h5 : C++ : add options 'w', 'r', 'a' for C++ file constructor to mimic Python
- block gf : clean the hdf5 format and save/load difference between Python and C++
- rewrite gf_base_op test to eliminate the h5 file (big).

+++ triqs/gfs/meshes/linear.hpp
@@ -89,7 +89,7 @@ namespace gfs {
    h5_write(gr, "domain", m.domain());
    h5_write(gr, "min", m.xmin);
    h5_write(gr, "max", m.xmax);
-   h5_write(gr, "size", m.size());
+   h5_write(gr, "size", long(m.size()));
   }
 
   /// Read from HDF5
