+++ triqs/gfs/meshes/matsubara_freq.hpp
@@ -146,7 +146,7 @@ namespace gfs {
   friend void h5_write(h5::group fg, std::string subgroup_name, gf_mesh const &m) {
    h5::group gr = fg.create_group(subgroup_name);
    h5_write(gr, "domain", m.domain());
-   h5_write(gr, "size", m.size());
+   h5_write(gr, "size", long(m.size()));
    h5_write(gr, "start_at_0", (m._positive_only?1:0));
   }
 
@@ -154,10 +154,11 @@ namespace gfs {
   friend void h5_read(h5::group fg, std::string subgroup_name, gf_mesh &m) {
    h5::group gr = fg.open_group(subgroup_name);
    typename gf_mesh::domain_t dom;
-   int L;
+   long L;
    int s = 1;
    h5_read(gr, "domain", dom);
    h5_read(gr, "size", L);
+   // backward compatibility : older file do not have this flags, default is true. 
    if (gr.has_key("start_at_0")) h5_read(gr, "start_at_0", s);
    m = gf_mesh{std::move(dom), L, (s==1)};
   }
